planner: split index test (#46780)
close pingcap/tidb#44940
