+++ cmd/mirror/mirror.go
@@ -197,7 +197,7 @@ func downloadZips(
 	cmd := exec.Command(gobin, downloadArgs...)
 	cmd.Dir = tmpdir
 	env := os.Environ()
-	env = append(env, fmt.Sprintf("GOPROXY=%s", "https://proxy.golang.org,direct"))
+	env = append(env, fmt.Sprintf("GOPROXY=%s", "https://mirrors.aliyun.com/goproxy/,https://proxy.golang.org,direct"))
 	env = append(env, fmt.Sprintf("GOSUMDB=%s", "sum.golang.org"))
 	cmd.Env = env
 	jsonBytes, err := cmd.Output()
@@ -228,7 +228,7 @@ func listAllModules(tmpdir string) (map[string]listedModule, error) {
 	cmd := exec.Command(gobin, "list", "-mod=readonly", "-m", "-json", "all")
 	cmd.Dir = tmpdir
 	env := os.Environ()
-	env = append(env, fmt.Sprintf("GOPROXY=%s", "https://proxy.golang.org,direct"))
+	env = append(env, fmt.Sprintf("GOPROXY=%s", "https://mirrors.aliyun.com/goproxy/,https://proxy.golang.org,direct"))
 	env = append(env, fmt.Sprintf("GOSUMDB=%s", "sum.golang.org"))
 	cmd.Env = env
 	jsonBytes, err := cmd.Output()
